+++ pyproject.toml
@@ -16,6 +16,9 @@ azure-core = "^1.33.0"
 validators = ">=0.35.0"
 data-platform-helpers = ">=0.1.7"
 charm-refresh = "^3.1.0.0"
+charmlibs-interfaces-tls-certificates = ">=1.0"
+charmlibs-snap = "^1.0.1"
+charmlibs-systemd = "^1.0.0"
 
 [tool.poetry.requires-plugins]
 poetry-plugin-export = ">=1.8"
