+++ src/events/tls.py
@@ -7,7 +7,7 @@
 import logging
 from typing import TYPE_CHECKING
 
-from charms.tls_certificates_interface.v4.tls_certificates import (
+from charmlibs.interfaces.tls_certificates import (
     CertificateAvailableEvent,
     CertificateRequestAttributes,
     TLSCertificatesRequiresV4,
